+++ rs/types/types/src/crypto/hash/tests.rs
@@ -1053,14 +1053,13 @@ mod crypto_hash_stability {
     fn canister_http_response_stability() {
         let data = CanisterHttpResponse {
             id: CanisterHttpRequestId::from(42),
-            timeout: UNIX_EPOCH,
             canister_id: ic_base_types::CanisterId::from_u64(42),
             content: CanisterHttpResponseContent::Success(vec![0x42; 16]),
         };
         let hash = crypto_hash(&data);
         assert_eq!(
             hex::encode(hash.get_ref().0.as_slice()),
-            "48f996c026833b51d8e34775ee8ca1354abe0b4fd29d9795290e3747ad477650",
+            "86afd80c0bbb31b1776437bee06ff99329e91b53a5afea4c46003874d3e9bf3d",
             "Hash of CanisterHttpResponse changed"
         );
     }
@@ -1070,7 +1069,6 @@ mod crypto_hash_stability {
     fn canister_http_response_metadata_stability() {
         let data = CanisterHttpResponseMetadata {
             id: CallbackId::from(42),
-            timeout: UNIX_EPOCH,
             content_hash: test_crypto_hash_of(0x42),
             content_size: 0,
             registry_version: RegistryVersion::from(1),
@@ -1079,7 +1077,7 @@ mod crypto_hash_stability {
         let hash = crypto_hash(&data);
         assert_eq!(
             hex::encode(hash.get_ref().0.as_slice()),
-            "be14ad8b8f7fe064b74586e784c669ddc0878c700076ba8bd186cfbd42701bfd",
+            "7c32c9aca0291c4440de5a0c7b370c6b977fdc9efff58847f8c8db6bb83f1f17",
             "Hash of CanisterHttpResponseMetadata changed"
         );
     }
@@ -1089,7 +1087,6 @@ mod crypto_hash_stability {
     fn canister_http_response_share_stability() {
         let metadata = CanisterHttpResponseMetadata {
             id: CallbackId::from(42),
-            timeout: UNIX_EPOCH,
             content_hash: test_crypto_hash_of(0x42),
             content_size: 0,
             registry_version: RegistryVersion::from(1),
@@ -1105,7 +1102,7 @@ mod crypto_hash_stability {
         let hash = crypto_hash(&data);
         assert_eq!(
             hex::encode(hash.get_ref().0.as_slice()),
-            "bfc544c49bfab42d9d2d153a9e840c68fd8b49a3ce70f361b1646e89d360184d",
+            "9adaec0a8b383039b5723e2dd4c5b27c545bcc43c9492f1037633ef725aafc73",
             "Hash of CanisterHttpResponseShare changed"
         );
     }